Green hue (vk/86b5-green-hue)
Cards with a counter above 5m have a green hue. But it appears that once a card got it's green hue it won't disappear even when the counter reached zero and even when the website was refreshed. Look for the countdown counter for cards and try to diagnose the issue