Card mix up their contents (vk/e4ce-card-mix-up-thei)
Currently the cards can completely mix up their context. New on-going cards output will show when viewing older cards. Real time updates can show in older card when newer ones are running. There is a lot of mix up in terms of contents. I'm assuming it has to do with memoization. We need to do it on per execution basis