Green hue timer not always propagates properly (vk/d66e-green-hue-timer)
Right now a green/orange hue is added to tasks that were recently touched (last 5 minutes), however when the timer runs out, it shows properly that the card is expired WHEN it's currently focused one. However when changing cards it can go back to appearing green even though the timer is actually at 0. Can you hunt down the issue?