Mobile diff bar misaligned (vk/8fb7-mobile-diff-bar)
When displaying Diff on mobile the bar that summarizes the changes is slightly below the top, while the diffs themselves start from the very top, which makes the summary bar cover the actual diff and display in the middle of the diff itself. Find the positioning issue. Put the bar on top and the diffs right below it